% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_stops.R
\name{.dubois_default_tx_counties}
\alias{.dubois_default_tx_counties}
\title{A fixed set of real Texas county FIPS codes used as the default sampling
frame for \code{\link[=simulate_stops]{simulate_stops()}} (Harris, Dallas, Travis, Bexar, Tarrant, El
Paso, Collin, Denton, Fort Bend, Hidalgo).}
\usage{
.dubois_default_tx_counties
}
\description{
A fixed set of real Texas county FIPS codes used as the default sampling
frame for \code{\link[=simulate_stops]{simulate_stops()}} (Harris, Dallas, Travis, Bexar, Tarrant, El
Paso, Collin, Denton, Fort Bend, Hidalgo).
}
\keyword{internal}
