% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veil_of_darkness_charts.R
\name{plot_county_search_disparity}
\alias{plot_county_search_disparity}
\title{Plot county-level search-rate disparity (chart 3)}
\usage{
plot_county_search_disparity(county_search_disparity, min_n = 30)
}
\arguments{
\item{county_search_disparity}{Output of
\code{\link[=summarize_county_search_disparity]{summarize_county_search_disparity()}}.}

\item{min_n}{Minimum \code{n_searches_black}/\code{n_searches_white} for a county
to be plotted. Default \code{30}.}
}
\value{
A \code{ggplot}.
}
\description{
Plot county-level search-rate disparity (chart 3)
}
