% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veil_of_darkness_charts.R
\name{plot_county_vod_disparity}
\alias{plot_county_vod_disparity}
\title{Plot the county-level Veil of Darkness ratio (chart 1)}
\usage{
plot_county_vod_disparity(county_vod_disparity, min_n = 30)
}
\arguments{
\item{county_vod_disparity}{Output of \code{\link[=summarize_county_vod_disparity]{summarize_county_vod_disparity()}}.}

\item{min_n}{Minimum \code{total_n} for a county to be plotted. Default \code{30}.}
}
\value{
A \code{ggplot}.
}
\description{
Plot the county-level Veil of Darkness ratio (chart 1)
}
