% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veil_of_darkness_charts.R
\name{plot_statewide_vod}
\alias{plot_statewide_vod}
\title{Plot statewide racial composition of stops, before vs. after dark (chart 2)}
\usage{
plot_statewide_vod(statewide_vod, race_col = "subject_race")
}
\arguments{
\item{statewide_vod}{Output of \code{\link[=summarize_statewide_vod]{summarize_statewide_vod()}}.}

\item{race_col}{String race column name in \code{statewide_vod}. Default
\code{"subject_race"}.}
}
\value{
A \code{ggplot}.
}
\description{
Plot statewide racial composition of stops, before vs. after dark (chart 2)
}
